+++ src/apps/cronus-finance/index.ts
@@ -1,3 +1,0 @@
-export { CRONUS_FINANCE_DEFINITION, CronusFinanceAppDefinition } from './cronus-finance.definition';
-export { CronusFinanceAppModule } from './cronus-finance.module';
-export { CronusFinanceContractFactory } from './contracts';