+++ src/apps/evmoswap/index.ts
@@ -1,3 +1,0 @@
-export { EVMOSWAP_DEFINITION, EvmoswapAppDefinition } from './evmoswap.definition';
-export { EvmoswapAppModule } from './evmoswap.module';
-export { EvmoswapContractFactory } from './contracts';